requirements.txt
# Backend
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
python-multipart>=0.0.6

# Frontend
streamlit>=1.29.0,<2.0.0

# Vector DB
chromadb>=0.5.0,<1.0.0

# LLM & Embeddings
openai>=1.10.0,<2.0.0

# Document Processing
PyMuPDF>=1.23.0
python-docx>=1.1.0,<2.0.0

# Search
rank-bm25>=0.2.2,<1.0.0

# Data Validation
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# Configuration
python-dotenv>=1.0.0,<2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.25.0
